+++ rubin_sim/moving_objects/base_obs.py
