+++ rubin_sim/moving_objects/direct_obs.py
@@ -2,7 +2,7 @@ import logging
 import numpy as np
 import datetime
 
-from .baseObs import BaseObs
+from .base_obs import BaseObs
 
 __all__ = ["DirectObs"]
