+++ rubin_sim/moving_objects/make_lsst_obs.py
