+++ tests/moving_objects/test_Orbits.py
@@ -3,7 +3,7 @@ import os
 import numpy as np
 import pandas as pd
 from pandas.testing import assert_frame_equal
-from rubin_sim.movingObjects import Orbits
+from rubin_sim.moving_objects import Orbits
 from rubin_sim.data import get_data_dir
